pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Felipe Saadi <felipesaadisiegert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
langchainhub = "^0.1.21"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.9"
langchain-anthropic = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
